[Refactor/TFLITE] Refactor tflite tensor-filter subplugin to remove .c/.cc wrapper
This commit refactors tflite tensor-filter subplugin to remove .c/.cc
wrapper.
