--- ext/nnstreamer/tensor_filter/tensor_filter_tensorflow_lite.c
@@ -1,242 +1,0 @@
-/**
- * GStreamer Tensor_Filter, Tensorflow-Lite Module
- * Copyright (C) 2018 MyungJoo Ham <myungjoo.ham@samsung.com>
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Library General Public
- * License as published by the Free Software Foundation; either
- * version 2 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Library General Public License for more details.
- *
- */
-/**
- * @file	tensor_filter_tensorflow_lite.c
- * @date	24 May 2018
- * @brief	Tensorflow-lite module for tensor_filter gstreamer plugin
- * @see		http://github.com/nnsuite/nnstreamer
- * @author	MyungJoo Ham <myungjoo.ham@samsung.com>
- * @bug		No known bugs except for NYI items
- *
- * This is the per-NN-framework plugin (tensorflow-lite) for tensor_filter.
- * Fill in "GstTensorFilterFramework" for tensor_filter.h/c
- *
- */
-
-#include <glib.h>
-#include <string.h>
-
-#include "tensor_filter_tensorflow_lite_core.h"
-
-void init_filter_tflite (void) __attribute__ ((constructor));
-void fini_filter_tflite (void) __attribute__ ((destructor));
-
-/**
- * @brief internal data of tensorflow lite
- */
-struct _Tflite_data
-{
-  void *tflite_private_data;
-};
-typedef struct _Tflite_data tflite_data;
-
-/**
- * @brief Free privateData and move on.
- */
-static void
-tflite_close (const GstTensorFilterProperties * prop, void **private_data)
-{
-  tflite_data *tf;
-  tf = *private_data;
-  tflite_core_delete (tf->tflite_private_data);
-  g_free (tf);
-  *private_data = NULL;
-}
-
-/**
- * @brief Load tensorflow lite modelfile
- * @param prop property of tensor_filter instance
- * @param private_data : tensorflow lite plugin's private data
- * @return 0 if successfully loaded. 1 if skipped (already loaded).
- *        -1 if the object construction is failed.
- *        -2 if the object initialization if failed
- */
-static int
-tflite_loadModelFile (const GstTensorFilterProperties * prop,
-    void **private_data)
-{
-  tflite_data *tf;
-  int err;
-
-  if (*private_data != NULL) {
-    /** @todo : Check the integrity of filter->data and filter->model_files, nnfw */
-    tf = *private_data;
-
-    if (tflite_core_compareModelPath (tf->tflite_private_data, prop->model_files[0]))
-      return 1;
-
-    tflite_close (prop, private_data);
-  }
-
-  tf = *private_data = g_new0 (tflite_data, 1);
-  if (tf == NULL) {
-    g_printerr ("Failed to allocate memory for filter subplugin.");
-    return -1;
-  }
-
-  tf->tflite_private_data = tflite_core_new (prop->model_files[0], prop->accl_str);
-  if (tf->tflite_private_data == NULL) {
-    g_printerr ("failed to create the object: Tensorflow-lite");
-    err = -1;
-    goto err_free;
-  }
-
-  if (tflite_core_init (tf->tflite_private_data) != 0) {
-    g_printerr ("failed to initialize the object: Tensorflow-lite");
-    tflite_core_delete (tf->tflite_private_data);
-    err = -2;
-    goto err_free;
-  }
-
-  return 0;
-
-err_free:
-  g_free (*private_data);
-  *private_data = NULL;
-
-  return err;
-}
-
-/**
- * @brief The open callback for GstTensorFilterFramework. Called before anything else
- * @param prop property of tensor_filter instance
- * @param private_data : tensorflow lite plugin's private data
- */
-static int
-tflite_open (const GstTensorFilterProperties * prop, void **private_data)
-{
-  int ret = tflite_loadModelFile (prop, private_data);
-  g_assert (ret >= 0);       /** This must be called only once */
-  return ret;
-}
-
-/**
- * @brief The mandatory callback for GstTensorFilterFramework
- * @param prop property of tensor_filter instance
- * @param private_data : tensorflow lite plugin's private data
- * @param[in] input The array of input tensors
- * @param[out] output The array of output tensors
- * @return 0 if OK. non-zero if error.
- */
-static int
-tflite_invoke (const GstTensorFilterProperties * prop, void **private_data,
-    const GstTensorMemory * input, GstTensorMemory * output)
-{
-  int retval;
-  tflite_data *tf;
-  tf = *private_data;
-  g_assert (*private_data);
-  retval = tflite_core_invoke (tf->tflite_private_data, input, output);
-  g_assert (retval == 0);
-  return retval;
-}
-
-/**
- * @brief The optional callback for GstTensorFilterFramework
- * @param prop property of tensor_filter instance
- * @param private_data : tensorflow lite plugin's private data
- * @param[out] info The dimesions and types of input tensors
- */
-static int
-tflite_getInputDim (const GstTensorFilterProperties * prop, void **private_data,
-    GstTensorsInfo * info)
-{
-  tflite_data *tf;
-  tf = *private_data;
-  g_assert (*private_data);
-  return tflite_core_getInputDim (tf->tflite_private_data, info);
-}
-
-/**
- * @brief The optional callback for GstTensorFilterFramework
- * @param prop property of tensor_filter instance
- * @param private_data : tensorflow lite plugin's private data
- * @param[out] info The dimesions and types of output tensors
- */
-static int
-tflite_getOutputDim (const GstTensorFilterProperties * prop,
-    void **private_data, GstTensorsInfo * info)
-{
-  tflite_data *tf;
-  tf = *private_data;
-  g_assert (*private_data);
-  return tflite_core_getOutputDim (tf->tflite_private_data, info);
-}
-
-/**
- * @brief The optional callback for GstTensorFilterFramework
- * @param prop property of tensor_filter instance
- * @param private_data : tensorflow lite plugin's private data
- * @param in_info The dimesions and types of input tensors
- * @param[out] out_info The dimesions and types of output tensors
- */
-static int
-tflite_setInputDim (const GstTensorFilterProperties * prop, void **private_data,
-    const GstTensorsInfo * in_info, GstTensorsInfo * out_info)
-{
-  tflite_data *tf;
-  tf = *private_data;
-  g_assert (*private_data);
-  return tflite_core_setInputDim (tf->tflite_private_data, in_info, out_info);
-}
-
-/**
- * @brief The optional callback for GstTensorFilterFramework
- * @param prop property of tensor_filter instance
- * @param private_data : tensorflow lite plugin's private data
- * @return 0 if OK. non-zero if error.
- */
-static int
-tflite_reloadModel (const GstTensorFilterProperties * prop, void **private_data)
-{
-  tflite_data *tf;
-  tf = *private_data;
-  g_assert (*private_data);
-
-  if (prop->num_models != 1)
-    return -1;
-
-  return tflite_core_reloadModel (tf->tflite_private_data, prop->model_files[0]);
-}
-
-static gchar filter_subplugin_tensorflow_lite[] = "tensorflow-lite";
-
-static GstTensorFilterFramework NNS_support_tensorflow_lite = {
-  .name = filter_subplugin_tensorflow_lite,
-  .allow_in_place = FALSE,      /** @todo: support this to optimize performance later. */
-  .allocate_in_invoke = FALSE,
-  .invoke_NN = tflite_invoke,
-  .getInputDimension = tflite_getInputDim,
-  .getOutputDimension = tflite_getOutputDim,
-  .setInputDimension = tflite_setInputDim,
-  .reloadModel = tflite_reloadModel,
-  .open = tflite_open,
-  .close = tflite_close,
-};
-
-/** @brief Initialize this object for tensor_filter subplugin runtime register */
-void
-init_filter_tflite (void)
-{
-  nnstreamer_filter_probe (&NNS_support_tensorflow_lite);
-}
-
-/** @brief Destruct the subplugin */
-void
-fini_filter_tflite (void)
-{
-  nnstreamer_filter_exit (NNS_support_tensorflow_lite.name);
-}
